script tag 적용, header / footer 적용 -> user

+++ Vortex/src/main/java/vortex/application/access/web/ActionController.java
@@ -24,7 +24,7 @@ public class ActionController extends ApplicationController {
 	@RequestMapping("/group/list.do")
 	public ModelAndView getGroups(HttpServletRequest hreq) {
 		DataObject req = request(hreq);
-		req.set("start", req.number("start"))
+		req.set("start", req.number("start").intValue())
 		   .set("fetch", properties.getInt("fetch"));
 		return modelAndView(!req.bool("ajax") ? "action/group/list" : "jsonView", actionService.getGroups(req)); 
 	}
