+++ Vortex/src/main/java/vortex/application/group/web/GroupController.java
@@ -21,7 +21,7 @@ public class GroupController extends ApplicationController {
 	@RequestMapping("/list.do")
 	public ModelAndView getGroups(HttpServletRequest hreq) {
 		DataObject req = request(hreq);
-		req.set("start", req.number("start"))
+		req.set("start", req.number("start").intValue())
 		   .set("fetch", properties.getInt("fetch"));
 		return modelAndView(!req.bool("ajax") ? "group/list" : "jsonView", groupService.getGroups(req)); 
 	}
