+++ Vortex/src/main/java/vortex/application/user/web/UserController.java
@@ -21,7 +21,7 @@ public class UserController extends ApplicationController {
 	@RequestMapping("/list.do")
 	public ModelAndView search(HttpServletRequest hreq) {
 		DataObject req = request(hreq);
-		req.set("start", req.number("start"))
+		req.set("start", req.number("start").intValue())
 		   .set("fetch", properties.getInt("fetch"));
 		return modelAndView(!req.bool("ajax") ? "user/list" : "jsonView", userService.search(req));
 	}
